7 - WebClient implementation failed rs

+++ src/main/java/com/portfolio/astrology/model/P11.java
@@ -20,10 +20,10 @@ public class P11 {
     String name;
 
     @Column(nullable = false, length = 100)
-    Long longitude;//": 92.15322965365729
+    Float longitude;//": 92.15322965365729
 
     @Column(nullable = false, length = 100)
-    Long longitudeSpeed;//": 0.9535995114940861
+    Float longitudeSpeed;//": 0.9535995114940861
 
 
 }
